refactor(filter-match): update logic and tests
It should be private and return an error hint with a type string. In case the value does not match, it should return "is invalid", otherwise, if an error occurs while parsing the regular expression, it should return "has an invalid rule". There is no need to check the data type of both the regular expression and the passed value because calling the reflection.String() always returns a string value without causing an error